secondary index: avoid special characters in default index names
In CQL, table names are limited to alphanumeric characters (letters, numbers
and underscores), but column names don't have such a limitation. When we
create a secondary index, its default name is constructed from the column
name - so can contain problematic characters. It can include the "/"
character. The problem is that the index name is then used, like a table
name, to create a directory with that name.

The test included in this patch demonstrates that before this patch, this
can be misused to create subdirectories anywhere in the filesystem, or to
crash Scylla when it fails to create a directory (which it considers an
unrecoverable I/O error).

In this patch we do what Cassandra does - remove all non-alphanumeric
characters from the indexed column name before constructing the default
index name. In the included test we verify that the constructed index
name is the same as in Cassandra, which is useful to know (e.g., because
knowing the index name is needed to DROP the index).

Also, this patch adds a second line of defense against the security problem
described above: It is now an error to create a schema with a slash or
null (the two characters not allowed in Unix filenames) in the keyspace
or table names. So if the first line of defense (CQL checking the validity
of its commands) fails, we'll have that second line of defense. I verified
that if I revert the default-index-name fix, the second line of defense
kicks in, and the index creation is aborted and cannot create files in
the wrong place to crash Scylla.

Fixes #3403

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20220221173445.939378-1-nyh@scylladb.com>