make schema for index-induced-mv
Indexes over collections use materialized views. Supposing that we're
deling with global indexes, and that pk, ck were the partition and
clustering keys of the base table, the schema of the materialized view
looks like this:
new_pk=coll_value, new_ck=concat(idx_token, pk, ck)

where coll_value is the computed value from the collection, be it a key
from the collection, a value from the collection, or the tuple conaining
both.